+++ test/test.ml
@@ -1,4 +1,6 @@
 let () =
   let path = Sys.getenv_opt "REPORT_PATH" in
-  let report = Junit.make [ OidcParameters.suite; Jwt.suite; Jwks.suite ] in
+  let report =
+    Junit.make [ OidcParameters.suite; Jwt.suite; Jwks.suite; Token.suite ]
+  in
   match path with Some path -> Junit.to_file report path | None -> ()